Settlement issue

See merge request mislbd-ng/services-ng/ababil-foreign-remittance!133

+++ foreign-remittance-impl/src/main/java/com/mislbd/ababil/foreignremittance/command/handler/TransactionReconcileCommandHandlerAggregate.java
@@ -152,7 +152,7 @@ public class TransactionReconcileCommandHandlerAggregate {
   private ReconcileTxnLogEntity prepareReconcileTransactionLog(
       Command<?> command,
       String accountNumber,
-      BigDecimal globalTxnNumber,
+      Long globalTxnNumber,
       LocalDate txnDate,
       String cause) {
     ReconcileTxnLogEntity entity = new ReconcileTxnLogEntity();
